Make sdk/ readable at a glance
The package listing gave a newcomer no signal about what mattered: four
private modules sat among the public ones, and nothing said which file served
which concern.

Measuring the package first was what made the shape obvious. Geometry is 54%
of it and verification another 25%; structure, physics and appearance together
are 13%. That is why grouping by physics/appearance never worked -- those
threads barely exist as code. They are fields on one class and two small
modules, while the bulk is a geometry library and a test framework.

So the machinery recedes and the map gets written down:

- _section_loft and _shell_partition join _mesh/. They are geometry internals;
  mesh.py already imported them alongside the rest of the engine.
- _mass_solver folds into mass.py. It already imported it, so there was no
  cycle, and "where does a part's mass come from" now has one file.
- The package docstring names the five threads and which module owns each,
  and says plainly that Material spans physics and appearance on purpose --
  which a physics/ folder would have had to lie about.

The agent's surface is unchanged and checked, not assumed: 50 names from
mini_articraft.sdk and 35 from .mesh, none added, none missing, all
importable.
