+++ src/mini_articraft/sdk/_mesh/shell_partition.py
