+++ src/mini_articraft/sdk/export.py
@@ -28,9 +28,9 @@ from pxr import (  # pyright: ignore[reportAttributeAccessIssue]
 
 from mini_articraft.sdk import ambientcg
 from mini_articraft.sdk._collision import MeshCollisionKernel, _rpy_matrix
-from mini_articraft.sdk._mass_solver import ResolvedMass, resolve_mass
 from mini_articraft.sdk._mesh.core import MeshGeometry, geometry_to_trimesh
 from mini_articraft.sdk.joints import Articulation, ArticulationType, MotionLimits
+from mini_articraft.sdk.mass import ResolvedMass, resolve_mass
 from mini_articraft.sdk.materials import Material, is_library_material
 from mini_articraft.sdk.object import ArticulatedObject, Geometry
 from mini_articraft.sdk.testing import DEFAULT_MESH_TOLERANCE
